Add Invite Users option to user menu
