+++ ui/apps/platform/src/Containers/MainPage/Header/UserMenu.tsx
@@ -1,0 +1,115 @@
+import React, { useState, CSSProperties } from 'react';
+import PropTypes from 'prop-types';
+import { withRouter } from 'react-router-dom';
+import { connect } from 'react-redux';
+import { createStructuredSelector } from 'reselect';
+import initials from 'initials';
+import { Dropdown, DropdownItem, DropdownSeparator, DropdownToggle } from '@patternfly/react-core';
+
+import usePermissions from 'hooks/usePermissions';
+import { selectors } from 'reducers';
+import { actions as authActions } from 'reducers/auth';
+import { userBasePath } from 'routePaths';
+import User from 'utils/User';
+
+const userMenuStyleConstant = {
+    '--pf-u-min-width--MinWidth': '20ch',
+    pointerEvents: 'none',
+} as CSSProperties;
+
+function UserMenu({ logout, userData }) {
+    const [isOpen, setIsOpen] = useState(false);
+    const { hasReadWriteAccess } = usePermissions();
+    const hasWriteAccessForInviting = hasReadWriteAccess('Access');
+
+    function onSelect() {
+        setIsOpen(false);
+    }
+
+    const user = new User(userData);
+    const { email, name, roles } = user;
+
+    const displayName = email ? `${name as string} (${email})` : name;
+    const displayRoles = Array.isArray(roles) ? roles.map((role) => role.name).join(',') : '';
+
+    function testInviteClick() {
+        console.log('Invite users clicked');
+    }
+    const startOfUserMenu = [
+        <DropdownItem
+            key="user"
+            description={displayRoles}
+            className="pf-u-min-width"
+            style={userMenuStyleConstant}
+        >
+            {displayName}
+        </DropdownItem>,
+        <DropdownSeparator key="separator-1" />,
+        <DropdownItem key="profile" href={userBasePath}>
+            My profile
+        </DropdownItem>,
+    ];
+
+    const endOfUserMenu = [
+        <DropdownSeparator key="separator-2" />,
+        <DropdownItem key="logout" component="button" onClick={logout}>
+            Log out
+        </DropdownItem>,
+    ];
+
+    const inviteMenuItem = (
+        <DropdownItem key="open-invite" onClick={testInviteClick}>
+            Invite users
+        </DropdownItem>
+    );
+
+    const dropdownItems = hasWriteAccessForInviting
+        ? [...startOfUserMenu, inviteMenuItem, ...endOfUserMenu]
+        : [...startOfUserMenu, ...endOfUserMenu];
+
+    const toggle = (
+        <DropdownToggle aria-label="User menu" onToggle={setIsOpen} toggleIndicator={null}>
+            <span className="h-10 w-10 flex items-center justify-center leading-none text-xl border border-base-400 rounded-full">
+                {name ? initials(name) : '--'}
+            </span>
+        </DropdownToggle>
+    );
+
+    return (
+        <Dropdown
+            dropdownItems={dropdownItems}
+            isOpen={isOpen}
+            isPlain
+            onSelect={onSelect}
+            position="right"
+            toggle={toggle}
+        />
+    );
+}
+
+UserMenu.propTypes = {
+    logout: PropTypes.func.isRequired,
+    userData: PropTypes.shape({
+        userInfo: PropTypes.shape({
+            username: PropTypes.string,
+            friendlyName: PropTypes.string,
+            roles: PropTypes.arrayOf(
+                PropTypes.shape({
+                    name: PropTypes.string,
+                })
+            ),
+        }),
+        userAttributes: PropTypes.arrayOf(PropTypes.shape({})),
+    }).isRequired,
+};
+
+const mapStateToProps = createStructuredSelector({
+    userData: selectors.getCurrentUser,
+});
+
+const mapDispatchToProps = (dispatch) => ({
+    // eslint-disable-next-line @typescript-eslint/no-unsafe-return
+    logout: () => dispatch(authActions.logout()),
+});
+
+export default withRouter(connect(mapStateToProps, mapDispatchToProps)(UserMenu));
